abstract enemymodel
